+++ src/features/mypage/components/AchievementSection/MathContentInfo.tsx
