+++ src/features/mypage/components/AchievementSection/MathGoalInfo.tsx
