--- src/features/mypage/components/Sections/index.ts
@@ -1,1 +1,0 @@
-export { default as ProblemSettingSection } from './ProblemSettingSection';
